FarmWiserApp/requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.2.0
MarkupSafe==2.0.1
numpy==1.22.4
scikit-learn==0.24.2
scipy==1.6.3
sklearn==0.0
threadpoolctl==2.1.0
Werkzeug==2.0.1
xgboost==1.6.1
gradio==3.0.19
requests==2.28.0
pandas==1.4.2
seaborn==0.11.2
cfscrape
bs4
lxml==4.9.1
feedparser
newspaper3k
nltk